Make honeycomb agent read the "timestamp" field.
By default, honeycomb-agent uses the k8s timestamp instead of any passed
`timestamp` key in the JSON. This is incorrect for trace spans, which
require the timestamp to be the start time of the span, not the time it
is logged out (the end time).

By setting this field explicitly, the agent uses our value when it's
passed.