Honeycomb proxy
We want long-term storage of our logs; the intent here is to put them in
a coldline storage bucket (via stackdriver), which means first we needed
to get honeycomb events into stackdriver somehow.

This commit creates a proxy we can add as a sidecar onto our
honeycomb-agent deploy such that events sent to honeycomb _also_ get
logged to the sidecar container's stdout. (k8s yaml to create the
sidecar is a subsequent commit in this PR.)

I'd initially hoped that honeycomb-kubernetes-agent had a config option
to write to stdout in addition to sending, but it does not; the idea of
using an nginx proxy was suggested to me by Honeycomb folks, and it has
the upside of being easier than making and upstreaming changes to both
libhoney-go and honeycomb-kubernetes-agent. (The proxy approach is also
future proof against us introducing a different honeycomb client. We've
already done that with our honeycomb-postgres agent, though that's just
a hypothetical since we don't want to preserve our postgres logs
indefinitely.)

The README.md describes several of the implementation details; I'd like
to consider making this folder a public repo (after removing the API
key, of course).

Note: API Key belongs to a trial account belonging to ismith.

https://trello.com/c/UuqAo900/3008-long-term-storage-of-json-logs