When event handler is missing, mark event.
Previously, events that were destined for worker that didn't exist went
through the `Incomplete path of put_back, meaning they were marked as
"new" again and retried in 5 minutes, indefinitely.

This is causing operational trouble, as we have many, many events in
this state that are all getting retried every 5 minutes.

Instead, if the handler doesn't exist at all, mark the event as
"missing", essentially dropping it on the floor for now (we can have a
way to re-schedule these events in the future).

Note that 1) we still record the 404 and 2) this does not change
behavior for when the worker exists but the result is incomplete.