Sample stroller events
Why?
====

Per https://ui.honeycomb.io/teams/dark/usage, we're at about our daily
events target for the plan we're in; and 15.68% of our events in May
2020 were stroller events. (So this is costly.)

Those events don't seem to bring us much value; I can't remember
looking at these since stroller was new, and dstrelau can't recall
looking at them at all.

Implementation
==============
So, we sample - 1/100 events was picked somewhat arbitrarily (and is
aggressive!), but that'd
lead to stroller being ~1.57% of our events, which seems reasonable.

(235M for the first 3 weeks of May, is ~313M/month, 1/100th of that is
~3M/month out of a 1500M quota)

Alternatives
============
1. I bet we could condense some of these events - for instance, we have
an event for "pushing to stroller" and one for the response; and a
similar request/response pair for segment.

I decided that it wasn't worth the effort right now; just using static
sampling achieves our goals ("spending less on low-value events")
without requiring a bigger project.

https://trello.com/c/7dJdfP2l/3133-evaluate-stroller-logs-see-if-we-can-sample-and-or-remove-events