DarkInternal::getTableSizes
Adds a postgres query to get table sizes for all tables in our postgres
database, a DarkInternal function for that.

Intended use: honeycomb graph showing growth of db tables over time;
that'll let us set alerts on things like "hey, this table is way bigger
than we thought, is that a problem?"

See comments for details on example data.

Post-merge followup: add a cron to ops that runs this.

https://trello.com/c/LV8AOrUc/3164-log-postgres-table-sizes-so-we-have-a-honeycomb-graph-we-can-set-some-monitoring-on-hey-storedeventsv2-is-huge-is-the-gc-broken