Don't rollbar the #active-editor error, it's very high frequency (potentially)
See https://dark-inc.slack.com/archives/C012RAN8MCM/p1590599645011900 ;
we used 69k rollbars for this in one day! Our monthly quota is 100k,
with every 1k over that being $1, so this adds up fast.

It might be a thing to have a counter/ratelimit for rollbar or recover
somewhere clientside - "this client has reported this N times already,
don't report again" - but this is an easier fix.

https://trello.com/c/zSBMn2Ri/3159-stop-rollbaring-the-error-thats-blown-our-monthly-quota