Better fix for the client end of the AddTrust/Sectigo ssl bug
Previous work on this was:
https://github.com/darklang/dark/pull/2441 (remove the Addtrust root)
https://github.com/darklang/dark/pull/2445 (fix whitespace error in

This is not the ideal solution; one of the things we've learned in the
course of this (cite:
https://www.agwa.name/blog/post/fixing_the_addtrust_root_expiration) is
that gnutls is less robust than openssl. While removing the problematic
AddTrust root solved our immediate client-side problem, it does not
solve the general problem of gnutls not properly handling this type of
cert chaining.

Over the weekend, "replace gnutls with openssl" was one of the
approaches we considered, but we couldn't quite make it work, because
ocurl 0.8.2 depends on gnutls.

This PR solves that by ... upgrading to ocurl 0.9.1, which depends on
curl-conf, and thus allows us to use openssl.

Notes on testing:
- It works locally

Notes on verifying which library is installed
- `grep -r gnutls` shows libcurl4-gnutls-dev in
esy.lock/opam/conf-libcurl.1-opam, _BUT_ ` ldd
_build/default/backend/bin/server.exe | grep curl` confirms we're
linking against libcurl.so.4, not libcurl-gnutls.so.4
- `apt search libcurl` confirms that the installed libcurl package is "(OpenSSL flavor)"
- `dpkg -S` confirms that libcurl.so.4 comes from that package.

Trello: https://trello.com/c/pSHmm177/3167-follow-up-after-ssl-incident