Add convert-if-to-match command
This command will show when the cursor is within an if expression. If
the cursor is within multiple if expressions, the inner if expression is
converted to a match expression.

The command tries to be a bit smart by converting if expressions with a
binop condition to a match expression that has one of the binops
operands as the match condition. For example:
if val == 2 then
 "ok"
else
 "not ok"

will be converted to
match val
 2 -> "ok"
 _ -> "not ok"

The command defaults to having the full condition as the match condition
like this:
match val > 2
 true -> "ok"
 false -> "not ok"