Disable flaky exe_flow_fades test
We took a shot at adding an explicit timeout, but this test still fails
often, so removing it.

Honeycomb: https://ui.honeycomb.io/dark/datasets/integration-tests/result/dHp72neRgVN

Graph shows 3 integration tests that have failed in the last 14 days.
This test has failed 36 times out of 154 runs - 20%. The other two tests
fail much less often - 11 and 2 times in that same span - so 36 is an
outlier.

(This is a follow up to
https://trello.com/c/HoIkaNzv/3161-attempt-to-make-exeflowfades-test-more-reliable
)