First part of new garbage collection code
A few sections to this change

New query
=========
The old query expressed approximately "delete N=10k events that are (a)
older than a week, (b) not in the most recent ten in that set". (That
is, delete old stuff, but let's make sure we always have traces to
display.)

Doing a row-numbering in an inner query was making this execute very
slowly, as was trying to do it over the entire space of
stored_events_v2.

To fix this, we did two things:
- the "last_ten" subquery is now simpler; we don't need to number all
the rows, we don't even need them in order, we just need the last ten
- we do this for a handler, not a canvas (or all canvases)

The new query can be found in Stored_event.trim_events_for_handler.

Of course, since our overall goal is to GC all events, we have to wrap
this in two loops. The outer loop will (not in this commit, but later)
be "for each canvas in the db"; the inner loop is "for each handler on
this canvas".

This makes the query run much faster, and avoids locking the events
table for a lengthy period.

New runner
==========
We now have a garbage_collector.exe that takes an action ("count" or
"delete"), a canvas_id:string, and a limit:int, and garbage collects
stored_event_v2s for that canvas.

Goals:
We're not yet running this automatically, because:
- this is some new SQL, so I want to run it by hand so I can monitor
runtime, db load, etc
- we have a backlog of traces needing GC

Down the road we may either move this binary into a cron (possibly k8s
cron or just a loop in its own container), and likely wrap it in a loop
over all canvases, but for now, this script is the entry point.

Why limit:int? To satisfy the tuning/performance validation (see how
quickly it runs) and catch-up-with-backlog goals. Limit here is how many
traces (at most) we should attempt to delete in a particular query.
Prior to this PR, we hardcoded this at 10,000. Now it's easy to change.

New Db function, fetch_count
============================
I want, for investigation and development reasons, to be able to run the
query as either a DELETE (for actual garbage collecting) or as `SELECT
COUNT(*)`.  Db.delete returns an int, the number of rows affected; the
new Db.fetch_count does the same, but assumes the count will be the sole
field returned from the sql query.

Not done in this commit
=======================
We actually have three tables that store traces. stored_events_v2 is
covered in this commit; function_results_v2 and function_arguments are
not. Once I'm satisfied that stored_events_v2 garbage collection is
working as desired, I'll come back for function_results_v2 and
function_arguments.

Both function_results_v2 and function_arguments have a trace_id that
maps back to stored_events_v2, so GCing these may be simpler, using a
left join on stored_events_v2 to find "old" function_arguments (etc)
that do not have a corresponding stored_events_v2.

It might be worth considering a migration to add an fk constraint on
stored_events_v2 such that we can do a cascading delete (deleting from
stored_events_v2 deletes the function_* records). But that decision can
wait for this PR to land.

Intended next steps
===================
Merge and run this by-hand in-cluster; see how fast it is to run against
paulshen-animalcrossing, since that's a canvas with a lot of traces.

Logs will show both runtime and # of deleted traces.

Assuming that goes well, come back and add a loop over all canvases, and
run by hand again until we're out of the woods ("the gc has caught up").

In parallel with that catch up work, figure out how we want to run this
automatically. Given lack of error handling in cron and workers, I'm
thinking maybe we add a single-pod deploy to our infra that just runs
the gc script in a loop. Log to honeycomb, but sample heavily.