update garbage_collector.exe to loop over all canvases
Running in 'count' (not delete) mode across all canvases:
- took 29m16s (long, but tolerable - and I expect that to go down as db size does)
- found 1995755 rows to delete (2M, out of 85M)
- generated 113k traces (`cat gc.log | grep trace.trace_id | wc -l`),
  because we have 102k handlers worth of traces (note: not all traces
  have a corresponding handler - 404s, for instance, have no handler)

Unclear how long it'll take in delete mode; doing this per-canvas we saw
it take 3-4x as long. And even if we have to kill the script partway
through, that's still progress - each handler gets its own transaction,
so it's not all-or-nothing.

https://trello.com/c/UhpUHJcj/3186-fix-garbage-collection-of-old-traces