keep db and secrets injection in the same place
This slightly changes the previous implementation in exposing secrets to the exe & ana engines.

For handlers it was possible to inject dbs and secrets pretty early on in the fn call stack as input vars. But for user functions that isn't possible, we end up having to pass secrets 4-5 fn calls deep.

The discrepency between how handlers and fns work make up for a huge reason why ufns are so buggy. We want to keep their differences minimal.

Changes done in this solution are as follows:
- add secrets to exec_state
- create a helper function global_input_vars, which converts dbs and secrets from exec_state to input_vars
- the injections now happens in 2 places in ast.ml execute_ast, instead lf 3 places across multiple files.

Alt appraoch: I originally considered putting the injection inside AST.exec but that function is used all over the place and recurse on itself. It seems more prone to unintended consequences. So I moved the injection to right before it's caller to put it in the symtable before calling exec.