Make garbage collector run automatically instead of in a script on Ian's laptop
This PR contains a new deployment, a single-pod that just loops the gc
code. It logs to honeycomb, sampling at 1/10 events (we can tweak this
up or down as needed).

It also contains a new binary, garbage_collector_worker.exe, that yields
before each canvas-deletion query so the healthcheck thread has an
opportunity to respond.

NOTE: I believe running ANALYZE on the three relevant tables
(stored_events_v2, function_arguments, function_results_v2) is necessary
for our getAndLogTableSizes counts to be updated (and thus to see
results in honeycomb). I have not automated that here, that is still a
thing I intend to do manually for now.  I suspect ANALYZE gets run
automatically for us at intervals, but not sure when. This note is
acknowledging that it has not been done in this PR, but is on my radar.

https://trello.com/c/UhpUHJcj/3186-fix-garbage-collection-of-old-traces