Fix fullstory off-after-48hrs behavior
Follow up to #2573

As we thought might happen
(https://github.com/darklang/dark/pull/2573#discussion_r437801650), the
behavior in 2573 results in users we want to exclude (older than 48 hrs)
having _very brief_ sessions recorded. That's no good, because our quota
is per-session.

This PR fixes that by turning on the devMode flag for those users - see
doc at: https://www.npmjs.com/package/@fullstory/browser

"devMode - Set to true if you want to deactivate FullStory in your
development environment. When set to true, FullStory will shutdown
recording and all subsequent SDK method calls will be no-ops. At the
time init is called with devMode: true, a single event call will be sent
to FullStory to indicate that the SDK is in devMode; this is to help
trouble-shoot the case that the SDK was accidentally set to devMode:
true in a production environment. Additionally, any calls to SDK methods
will console.warn that FullStory is in devMode. Defaults to false."

So this is mostly the behavior we want; it is a little unfortunate that
there is the console.warn() log mentioned in that quote. I have
addressed that with a console log on startup that this is expected and
fine; see PR for a screenshot of the warning.

Once this has shipped, I'll coordinate with Victoria around testing.

https://trello.com/c/MD4bmVrK/3163-only-record-fullstory-for-users-whose-accounts-were-created-in-the-past-48-hours?menu=filter&filter=member:iansmith286