Try crlf attributes for paket.restore.targets
The Paket folks seem to not really know what to do with this file, but checking
it in seems the most likely correct thing. However, it seems to be regenerated
by the build, using crlf endings. This shows up in git status, breaking the
build.

So let's just make it all crlf