This switches to a less performant but more correct F# task implementation
Still faster than wrapping everything in a task, but slower (probably due to
more allocations) than what was the before. However, with the DTask version,
there was no type safety. Everything could be a task and you could have lists of
tasks. Rather, we'd like to get a proper dval out the back of it, and then be
able to do things to it.