Big change: reorganize and split Runtime types into ProgramTypes and RuntimeTypes
The goal here was to add function application into the Runtime, to be able to
remove the hack that is inject_param_and_execute. I wanted to add more options
to pipe, and to allow function values to exist and to be callable.

In most functional languages, there are names that refer to the symbol table,
and the symbol table can have function names in them. Rather than having names
in the symbol table (where they would be sorta mutable), add support for
references to immutably named functions.