Load static values lazily
Static values (those which are loaded when the process is initialized) cause
problems due to ordering (in which order are static processed - hard to detect)
as well as performance (they all have to be done, even if not used).

This makes them lazy. The intent right now is to ensure that all the
JsonConverters are registered before any JsonSerializer is used (the options are
immutable once they've been used, so we weren't able to add formatters during
initialization if serialization happened during static loading)