The redirect mechanism is no longer needed
We wanted to redirect some but not all. We didn't want to redirect the dev
server since it's only hosted on http: we did so with a config var.

We didn't want to redirect for healthchecks and solved this by moving the
healthchecks to their own server and putting the httpsredirect in the dark handler