Rewrite the Dockerfile to avoid file permissions
An issue that we're seeing is that things fail to build in VsCode because of
permissions. This redoes the dockerfile to make sure we use the `dark` user
wherever possible so that things aren't owned by root.

This adds all the mounts as directories so that they have the right
permissions when mounted.

Also did a bit of inlining