Get TaskOrValue properly working
The previous implementation broke in different ways - fixing one thing
would break the other. Eventually, I realized the issue: in
https://fsharpforfunandprofit.com/posts/computation-expressions-builder-part3/
they explain that the parameter of Combine, Run, and While are not
actually M<'t>, but should be the return type of Delay. Since I had
copied code from various sources and then hacked it til the type checker
allowed it, there were several bugs that came from the inconsistancies
of the various functions I had copied.

Obviously the docs don't mention this, that's for another day.