Attempt to avoid bash during Remote Container bootstrap
When using VSCode with the Remote Containers extension, the extension
bootstraps a container using a default alpine container. The alpine
container does not have bash.

When executing the `initializeCommand` script, the build fails when
trying to call bash. I speculate that it is trying to call bash because
there's a string in initializeCommand, but that it wouldn't if there was
instead an array - which doesn't need a shell to sort out.