Significant refactor of exceptions
This started as an attempt to get better output in rollbar. Because we were using a single exception type, all the exceptions reported looked like "Prelude+DarkException: an exception was thrown".

The exceptions were like that because that made sense in OCaml, but in F# it's more natural to have a different exception of each type. This also simplified a lot of the exception handling logic.

At the same time, I noticed a lot of functions that handled exceptions took metadata and a message in addition to the exception, and that they didn't always get propagated. This happened when there were centralized points via which many different exceptions of unknown types would go through. So this was a pattern where really I was reinventing innerExceptions. Since this is well-supported in .NET, I doubled down on that.

That also made it much simpler to support lastDitch exception handlers and pageable exceptions.