Attempt to fix exception in the CronChecker
The Cronchecker is supposed to wait until the DB is ready to start up. The exceptions
in Rollbar indicate that it is trying to run the crons before the DB is actually
ready for it, implying it got past the WaitForDB call.

This could be because the `return` was in the wrong place, executing after the first
iteration of the loop, as opposed to after the loop is over.