Add DeletedForever and fix delete behaviour
In the OCaml version, when we deleted we would also clear the cols used to search for a handler. But, in the F# version, I didn't do that by accident, causing issues where a deleted handler might be found. So this now deletes them properly.

When something is deleted forever, it's now actually deleted instead of just cleared.