Adjust BwdServer HSTS set-up, including tests
The implementation here is a bit of a hack; for
some reason, the constructs used in ApiServer.fs are
not working here. A follow-up commit is likely to clean
this up.

Note: 205/211 needed adjustment; the other 6
did not, as the requests do not hit our HTTP stack

Several (5) of the tests required careful adjustment
through vim because of their inconsistent line endings.