Don't use dictionaries to save values
This makes it expensive to add values, as we have to get a hashcode for them, which
for large values is expensive and may cause stackoverflows. Just use a list instead.
This does cause more values to be added, but we're going to switch this in a while to
a different implementation which will keep the values, so it's not important. Plus,
that's how the old version worked, so it can't be that bad.