Adjust tuple separator handling:
Previously, hitting , before _or_ after an existing tuple
separator would cause another element to be added. This
changes Fluid such that if we hit , here: (12|,34), we'll
just advance to (12,|34), while inserting a , in (12,|34)
will still cause a new element to be inserted, leaving us
with (12,|___,34).

This way, typing out (12,34) will result in the tuple being
complete, with no hanging blanks lying around