Add new match-evaluation logic to the interpreter
This is currently alongside the existing logic, in preparation for
fuzztests that should ensure that the two functions behave the same,
before committing to a migration to the new logic.

The new logic involves much less mutable state management, and is aimed
in preparation for a Tuple pattern which will require its sub-patterns
to all be matched in order to consider the full pattern matched.